+++ app/src/main/java/com/orange/ods/app/ui/components/buttons/ButtonsTextToggleGroup.kt
@@ -79,7 +79,7 @@ fun ButtonsTextToggleButtonsRow(customizationState: ButtonCustomizationState) {
                     parameters = {
                         list("textToggleButtons") {
                             textToggleButtons.take(toggleCount.value).forEach { item ->
-                                classInstance(OdsTextToggleButtonsRowItem::class.java) {
+                                classInstance<OdsTextToggleButtonsRowItem> {
                                     text(item.text)
                                     enabled(customizationState.isEnabled)
                                 }