+++ app/src/main/java/com/orange/ods/app/ui/components/buttons/icons/ButtonsIcon.kt
@@ -74,7 +74,7 @@ fun ButtonsIcon(customizationState: ButtonIconCustomizationState) {
                     name = OdsComposable.OdsIconButton.name,
                     exhaustiveParameters = false,
                     parameters = {
-                        classInstance("icon", OdsIconButtonIcon::class.java) {
+                        classInstance<OdsIconButtonIcon>("icon") {
                             painter()
                             contentDescription("")
                         }